separate application functionality from cross section classes
