+++ analysis/apps/cex_beam_quality_fits.py
@@ -14,7 +14,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from rich import print
-from python.analysis import Master, BeamParticleSelection, vector, Plots, cross_section, Fitting
+from python.analysis import Master, BeamParticleSelection, vector, Plots, cross_section, Fitting, Application
 
 def Fit_Vector(v : ak.Record, bins : int) -> tuple[dict, dict, dict, dict]:
     """ Gaussian fit to each component in the vector.
@@ -210,13 +210,13 @@ def main(args):
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description = "Computes Guassian fit paramters needed for the beam quality cuts in the beam particle selection.", formatter_class = argparse.RawDescriptionHelpFormatter)
 
-    cross_section.ApplicationArguments.Config(parser, True)
-    cross_section.ApplicationArguments.Processing(parser)
-    cross_section.ApplicationArguments.Output(parser)
-    cross_section.ApplicationArguments.Regen(parser)
+    Application.ApplicationArguments.Config(parser, True)
+    Application.ApplicationArguments.Processing(parser)
+    Application.ApplicationArguments.Output(parser)
+    Application.ApplicationArguments.Regen(parser)
 
     args = parser.parse_args()
 
-    args = cross_section.ApplicationArguments.ResolveArgs(args)
+    args = Application.ApplicationArguments.ResolveArgs(args)
     print(vars(args))
     main(args)
